Add scoped mesh properties
Adds `shell_mesh` and `solid_mesh` properties to the `Model`, which
are scoped to return only either shell or solid elements, respectively.

Adds the same properties `mesh`, `shell_mesh`, and `solid_mesh` to the
other objects, where applicable.

The property implementation is moved from 'model.py' to a new file
'_mesh.py'.
