+++ src/ansys/acp/core/_tree_objects/boolean_selection_rule.py
@@ -34,6 +34,7 @@ from ._grpc_helpers.property_helper import (
     grpc_data_property_read_only,
     mark_grpc_properties,
 )
+from ._mesh import mesh_property, shell_mesh_property
 from ._mesh_data import (
     ElementalData,
     NodalData,
@@ -119,6 +120,10 @@ class BooleanSelectionRule(CreatableTreeObject, IdTreeObject):
     )
 
     include_rule: ReadWriteProperty[bool, bool] = grpc_data_property("properties.include_rule_type")
+
+    mesh = mesh_property
+    shell_mesh = shell_mesh_property
+    # selection rules don't have solid mesh data
 
     elemental_data = elemental_data_property(BooleanSelectionRuleElementalData)
     nodal_data = nodal_data_property(BooleanSelectionRuleNodalData)
