+++ src/ansys/acp/core/_tree_objects/cutoff_selection_rule.py
@@ -34,6 +34,7 @@ from ._grpc_helpers.property_helper import (
     grpc_link_property,
     mark_grpc_properties,
 )
+from ._mesh import mesh_property, shell_mesh_property
 from ._mesh_data import (
     ElementalData,
     NodalData,
@@ -161,5 +162,8 @@ class CutoffSelectionRule(CreatableTreeObject, IdTreeObject):
     )
     ply_tapering: ReadWriteProperty[bool, bool] = grpc_data_property("properties.ply_tapering")
 
+    mesh = mesh_property
+    shell_mesh = shell_mesh_property
+    # selection rules don't have solid mesh data
     elemental_data = elemental_data_property(CutoffSelectionRuleElementalData)
     nodal_data = nodal_data_property(CutoffSelectionRuleNodalData)
