+++ src/ansys/acp/core/_tree_objects/parallel_selection_rule.py
@@ -35,6 +35,7 @@ from ._grpc_helpers.property_helper import (
     grpc_link_property,
     mark_grpc_properties,
 )
+from ._mesh import mesh_property, shell_mesh_property
 from ._mesh_data import (
     ElementalData,
     NodalData,
@@ -151,6 +152,10 @@ class ParallelSelectionRule(CreatableTreeObject, IdTreeObject):
         "properties.relative_rule_type"
     )
     include_rule: ReadWriteProperty[bool, bool] = grpc_data_property("properties.include_rule_type")
+
+    mesh = mesh_property
+    shell_mesh = shell_mesh_property
+    # selection rules don't have solid mesh data
 
     elemental_data = elemental_data_property(ParallelSelectionRuleElementalData)
     nodal_data = nodal_data_property(ParallelSelectionRuleNodalData)
