+++ src/ansys/acp/core/_tree_objects/tube_selection_rule.py
@@ -35,6 +35,7 @@ from ._grpc_helpers.property_helper import (
     grpc_link_property,
     mark_grpc_properties,
 )
+from ._mesh import mesh_property, shell_mesh_property
 from ._mesh_data import (
     ElementalData,
     NodalData,
@@ -159,5 +160,8 @@ class TubeSelectionRule(CreatableTreeObject, IdTreeObject):
         "properties.tail_extension"
     )
 
+    mesh = mesh_property
+    shell_mesh = shell_mesh_property
+    # selection rules don't have solid mesh data
     elemental_data = elemental_data_property(TubeSelectionRuleElementalData)
     nodal_data = nodal_data_property(TubeSelectionRuleNodalData)
