+++ src/ansys/acp/core/_tree_objects/variable_offset_selection_rule.py
@@ -38,6 +38,7 @@ from ._grpc_helpers.property_helper import (
     grpc_link_property,
     mark_grpc_properties,
 )
+from ._mesh import mesh_property, shell_mesh_property
 from ._mesh_data import (
     ElementalData,
     NodalData,
@@ -175,5 +176,8 @@ class VariableOffsetSelectionRule(CreatableTreeObject, IdTreeObject):
         "properties.distance_along_edge"
     )
 
+    mesh = mesh_property
+    shell_mesh = shell_mesh_property
+    # selection rules don't have solid mesh data
     elemental_data = elemental_data_property(VariableOffsetSelectionRuleElementalData)
     nodal_data = nodal_data_property(VariableOffsetSelectionRuleNodalData)
